+++ src/interactions/col_lam.c
@@ -16,19 +16,19 @@
 Term col_lam(Term col, Term lam) {
   interaction_count++;
   //printf("col_lam\n");
-  uint32_t col_loc = TERM_VAL(col);
-  uint32_t lam_loc = TERM_VAL(lam);
+  uint64_t col_loc = TERM_VAL(col);
+  uint64_t lam_loc = TERM_VAL(lam);
   uint8_t col_lab = TERM_LAB(col);
   uint8_t is_co0 = TERM_TAG(col) == CO0;
 
   Term bod = heap[lam_loc + 0];
 
   // Create new lambda variables
-  uint32_t lam0_loc = alloc(1);
-  uint32_t lam1_loc = alloc(1);
+  uint64_t lam0_loc = alloc(1);
+  uint64_t lam1_loc = alloc(1);
 
   // Create a superposition for lambda's variable
-  uint32_t sup_loc = alloc(2);
+  uint64_t sup_loc = alloc(2);
   heap[sup_loc + 0] = make_term(VAR, 0, lam0_loc);
   heap[sup_loc + 1] = make_term(VAR, 0, lam1_loc);
 
@@ -36,7 +36,7 @@ Term col_lam(Term col, Term lam) {
   heap[lam_loc] = make_sub(make_term(SUP, col_lab, sup_loc));
 
   // Create a collapser for lambda's body
-  uint32_t col_new_loc = alloc(1);
+  uint64_t col_new_loc = alloc(1);
   heap[col_new_loc] = bod;
 
   // Set up new lambda bodies
