+++ src/parse/term/all.c
@@ -3,7 +3,7 @@
 #include "../../memory.h"
 
 // Parse a pi type
-void parse_term_all(Parser* parser, uint32_t loc) {
+void parse_term_all(Parser* parser, uint64_t loc) {
   if (check_utf8(parser, 0xCE, 0xA0)) {
     consume_utf8(parser, 2);
   } else if (!consume(parser, "Π")) {
@@ -13,9 +13,9 @@ void parse_term_all(Parser* parser, uint32_t loc) {
   char* name = parse_name(parser);
   expect(parser, ":", "after name in pi type");
 
-  uint32_t all_node = alloc_term(2);
-  uint32_t inp_loc = all_node;
-  uint32_t out_loc = all_node + 1;
+  uint64_t all_node = alloc_term(2);
+  uint64_t inp_loc = all_node;
+  uint64_t out_loc = all_node + 1;
 
   parse_term(parser, inp_loc);
   expect(parser, ".", "after input type in pi type");
