+++ src/parse/term/efq.c
@@ -3,11 +3,11 @@
 #include "../../memory.h"
 
 // Parse an empty type elimination
-void parse_term_efq(Parser* parser, uint32_t loc) {
+void parse_term_efq(Parser* parser, uint64_t loc) {
   if (check_utf8(parser, 0xC2, 0xAC)) {
     consume_utf8(parser, 2);
 
-    uint32_t efq_node = alloc_term(1);
+    uint64_t efq_node = alloc_term(1);
     parse_term(parser, efq_node);
 
     store_term(loc, EFQ, 0, efq_node);
