+++ src/parse/term/rwt.c
@@ -3,12 +3,12 @@
 #include "../../memory.h"
 
 // Parse an equality elimination
-void parse_term_rwt(Parser* parser, uint32_t loc) {
+void parse_term_rwt(Parser* parser, uint64_t loc) {
   expect(parser, "%", "for equality elimination");
 
-  uint32_t rwt_node = alloc_term(2);
-  uint32_t eq_loc = rwt_node;
-  uint32_t bod_loc = rwt_node + 1;
+  uint64_t rwt_node = alloc_term(2);
+  uint64_t eq_loc = rwt_node;
+  uint64_t bod_loc = rwt_node + 1;
 
   parse_term(parser, eq_loc);
   expect(parser, ";", "after equality proof");
