+++ src/parse/term/set.c
@@ -3,7 +3,7 @@
 #include "../../memory.h"
 
 // Parse a universe type
-void parse_term_set(Parser* parser, uint32_t loc) {
+void parse_term_set(Parser* parser, uint64_t loc) {
   expect(parser, "*", "for universe type");
   store_term(loc, SET, 0, 0);
 }
