+++ src/parse/term/sup.c
@@ -3,15 +3,15 @@
 #include "../../memory.h"
 
 // Parse a superposition
-void parse_term_sup(Parser* parser, uint32_t loc) {
+void parse_term_sup(Parser* parser, uint64_t loc) {
   expect(parser, "&", "for superposition");
 
   uint8_t label = parse_uint(parser) & 3; // Ensure it fits in 2 bits
   expect(parser, "{", "after label in superposition");
 
-  uint32_t sup_node = alloc_term(2);
-  uint32_t lft_loc = sup_node;
-  uint32_t rgt_loc = sup_node + 1;
+  uint64_t sup_node = alloc_term(2);
+  uint64_t lft_loc = sup_node;
+  uint64_t rgt_loc = sup_node + 1;
 
   parse_term(parser, lft_loc);
   expect(parser, ",", "between terms in superposition");
